+++ app/javascript/engine/support/counterLayout.ts
@@ -14,9 +14,13 @@ import {
   targetSelectColor
 } from "../../utilities/graphics"
 import { canAssaultMove, canFire, canIntensiveFire, canMove, canReactionFire, canReactionIntensiveFire, canRout, canRush } from "../control/actionsAvailable"
+import { showClearObstacles, showEntrench } from "../control/assault"
+import { closeCombatCasualtyNeeded } from "../control/closeCombat"
+import { movementPastCost } from "../control/movement"
 import { stateType } from "../control/state/BaseState"
 import Counter from "../Counter"
 import Map from "../Map"
+import { gamePhaseType } from "./gamePhase"
 
 export function counterPath(counter: Counter, xOffset: number = 0, yOffset: number = 0): string {
   return baseCounterPath(counter.x + xOffset, counter.y + yOffset)
@@ -40,7 +44,7 @@ export function counterOutlineStyle(counter: Counter): SVGStyle {
   }
   if (counter.targetUF.selected) {
     return { fill: color, stroke: selectColor(), strokeWidth: 4 }
-  } else if (counter.hasUnit && counter.unit.targetSelected) {
+  } else if (counter.targetUF.targetSelected) {
     return { fill: color, stroke: targetSelectColor(), strokeWidth: 4 }
   } else if (counter.unit.dropSelected) {
     return { fill: color, stroke: dropSelectColor(), strokeWidth: 4 }
@@ -147,68 +151,110 @@ export function counterStatusLayout(counter: Counter): StatusLayout | boolean {
 export function counterActionButtons(
   map: Map, x: number, y: number, maxY: number, counter: Counter,
 ): ActionButtonLayout[] {
+  if (!map.game) { return [] }
   const rc: { x: number, color: string, text: string, tColor: string, action: string }[] = []
   const size = 24
   const boxHeight = 30
   const boxWidth = 40
-  const offset = 0
   let start = y
   if (y + boxHeight > maxY) {
     start = y - 176 - boxHeight
   }
-  if (map.game?.gameState?.type === stateType.Fire) {
-    if (counter.unit.targetSelected && map.game?.gameState?.type === stateType.Fire) {
-      rc.push({ x, color: actionGreen(), text: "Y", tColor: "#FFF", action: "fire_finish" })
-      rc.push({ x: x + boxWidth, color: counterRed(), text: "N", tColor: "#FFF", action: "cancel_action" })
-    }
-  } else if (map.game?.gameState?.type === stateType.Move) {
-    // if (map.game.moveState.path.length > 0) {
-    //   rc.push({ x, color: actionGreen(), text: "Y", tColor: "#FFF", action: "fire_finish" })
-    //   rc.push({ x: x + boxWidth, color: markerYellow(), text: "U", tColor: "#000", action: "undo_action" })
-    //   rc.push({ x: x + boxWidth, color: counterRed(), text: "N", tColor: "#FFF", action: "cancel_action" })
-    //   offset = -boxWidth * 1.5
-    // }
-  } else if (map.game?.gameState?.type === stateType.Assault) {
-    // if (map.game.assaultState.path.length > 0) {
-    //   rc.push({ x, color: actionGreen(), text: "Y", tColor: "#FFF", action: "fire_finish" })
-    //   rc.push({ x: x + boxWidth, color: counterRed(), text: "N", tColor: "#FFF", action: "cancel_action" })
-    //   offset = -boxWidth
-    // }
-  } else if (map.game?.gameState?.type === stateType.Rout) {
+  if (map.game?.phase === gamePhaseType.Deploy) {
     if (counter.unit.selected) {
-      rc.push({ x, color: counterRed(), text: "N", tColor: "#FFF", action: "cancel_action" })
+      rc.push({ x, color: markerYellow(), text: "U", tColor: "#000", action: "undeploy" })
     }
-  } else if (map.game?.gameState === undefined) {
+  } else if (map.game.phase === gamePhaseType.PrepRally) {
+    if (map.game.gameState?.type === stateType.Rally) {
+      if (counter.unit.selected) {
+        if (counter.unit.isBroken) {
+          rc.push({ x, color: actionGreen(), text: "R", tColor: "#FFF", action: "rally" })
+        } else {
+          rc.push({ x, color: actionGreen(), text: "R", tColor: "#FFF", action: "repair" })
+        }
+      }
+    }
+  } else if (map.game.phase === gamePhaseType.Main) {
+    if (map.game.gameState?.type === stateType.Fire) {
+      if (counter.unit.targetSelected && map.game?.gameState?.type === stateType.Fire) {
+        rc.push({ x, color: actionGreen(), text: "Y", tColor: "#FFF", action: "fire_finish" })
+        rc.push({ x: x + boxWidth, color: counterRed(), text: "N", tColor: "#FFF", action: "cancel_action" })
+      }
+    } else if (map.game.gameState?.type === stateType.Move) {
+      if (counter.unit.selected && counter.unit.canCarrySupport && counter.unit.smokeCapable &&
+          movementPastCost(map, counter.unit) < counter.unit.currentMovement) {
+        rc.push({ x, color: markerYellow(), text: "S", tColor: "#000", action: "move_smoke_toggle" })
+      }
+    } else if (map.game.gameState?.type === stateType.Assault) {
+      if (counter.unit.selected) {
+        if (showClearObstacles(map.game)) {
+          rc.push({ x, color: markerYellow(), text: "C", tColor: "#000", action: "assault_move_clear" })
+        }
+        const x2 = x + rc.length * boxWidth
+        if (showEntrench(map.game)) {
+          rc.push({ x: x2, color: markerYellow(), text: "D", tColor: "#000", action: "assault_move_entrench" })
+        }
+      }
+    } else if (map.game.gameState?.type === stateType.Rout) {
+      if (counter.unit.selected) {
+        rc.push({ x, color: counterRed(), text: "N", tColor: "#FFF", action: "cancel_action" })
+      }
+    } else if (map.game.gameState?.type === stateType.Reaction) {
+      if (counter.unit.selected) {
+        if (canReactionFire(counter.unit, map)) {
+          rc.push({ x, color: counterRed(), text: "F", tColor: "#FFF", action: "reaction_fire" })
+        } else if (canReactionIntensiveFire(counter.unit, map)) {
+          rc.push({ x, color: counterRed(), text: "F", tColor: "#FFF", action: "reaction_intensive_fire" })
+        }
+      }
+    } else if (map.game.gameState?.type === stateType.MoraleCheck) {
+      if (counter.unit.selected) {
+        rc.push({ x, color: markerYellow(), text: "M", tColor: "#000", action: "morale_check" })
+      }
+    } else if (map.game.gameState?.type === stateType.Breakdown) {
+      if (counter.unit.selected) {
+        rc.push({ x, color: markerYellow(), text: "B", tColor: "#000", action: "breakdown" })
+      }
+    } else if (map.game.gameState?.type === stateType.RoutCheck) {
+      if (counter.unit.selected) {
+        rc.push({ x, color: markerYellow(), text: "R", tColor: "#000", action: "rout_check" })
+      }
+    } else if (map.game.gameState === undefined) {
+      if (counter.unit.selected) {
+        if (canFire(counter.unit, map)) {
+          rc.push({ x, color: counterRed(), text: "F", tColor: "#FFF", action: "fire" })
+        } else if (canIntensiveFire(counter.unit, map)) {
+          rc.push({ x, color: counterRed(), text: "F", tColor: "#FFF", action: "intensive_fire" })
+        }
+        const x2 = x + rc.length * boxWidth
+        if (canMove(counter.unit, map)) {
+          rc.push({ x: x2, color: actionGreen(), text: "M", tColor: "#FFF", action: "move" })
+        } else if (canRush(counter.unit, map)) {
+          rc.push({ x: x2, color: actionGreen(), text: "M", tColor: "#FFF", action: "rush" })
+        }
+        const x3 = x + rc.length * boxWidth
+        if (canAssaultMove(counter.unit)) {
+          rc.push({ x: x3, color: actionBlue(), text: "A", tColor: "#FFF", action: "assault_move" })
+        }
+        const x4 = x + rc.length * boxWidth
+        if (canRout(counter.unit)) {
+          rc.push({ x: x4, color: markerYellow(), text: "R", tColor: "#000", action: "rout" })
+        }
+      }
+    }
+  } else if (map.game.phase === gamePhaseType.CleanupCloseCombat) {
+    if (counter.unit.selected && closeCombatCasualtyNeeded(map.game)) {
+      rc.push({ x, color: counterRed(), text: "H", tColor: "#FFF", action: "close_combat_reduce" })
+    }
+  } else if (map.game.phase === gamePhaseType.CleanupOverstack) {
     if (counter.unit.selected) {
-      if (canFire(counter.unit, map)) {
-        rc.push({ x, color: counterRed(), text: "F", tColor: "#FFF", action: "fire" })
-      } else if (canIntensiveFire(counter.unit, map)) {
-        rc.push({ x, color: counterRed(), text: "F", tColor: "#FFF", action: "intensive_fire" })
-      } else if (canReactionFire(counter.unit, map)) {
-        rc.push({ x, color: counterRed(), text: "F", tColor: "#FFF", action: "reaction_fire" })
-      } else if (canReactionIntensiveFire(counter.unit, map)) {
-        rc.push({ x, color: counterRed(), text: "F", tColor: "#FFF", action: "reaction_intensive_fire" })
-      }
-      const x2 = x + rc.length * (boxWidth + 0)
-      if (canMove(counter.unit, map)) {
-        rc.push({ x: x2, color: actionGreen(), text: "M", tColor: "#FFF", action: "move" })
-      } else if (canRush(counter.unit, map)) {
-        rc.push({ x: x2, color: actionGreen(), text: "M", tColor: "#FFF", action: "rush" })
-      }
-      const x3 = x + rc.length * (boxWidth + 0)
-      if (canAssaultMove(counter.unit)) {
-        rc.push({ x: x3, color: actionBlue(), text: "A", tColor: "#FFF", action: "assault_move" })
-      }
-      const x4 = x + rc.length * (boxWidth + 0)
-      if (canRout(counter.unit)) {
-        rc.push({ x: x4, color: markerYellow(), text: "R", tColor: "#000", action: "rout" })
-      }
+      rc.push({ x, color: counterRed(), text: "E", tColor: "#FFF", action: "overstack_reduce" })
     }
   }
   return rc.map(t => {
     return {
-      path: roundedRectangle(t.x - offset, start, boxWidth, boxHeight, 0), color: t.color, text: t.text,
-      tColor: t.tColor, size, tX: t.x + boxWidth/2 - offset, tY: start + boxHeight/2 + 6, action: t.action,
+      path: roundedRectangle(t.x, start, boxWidth, boxHeight, 0), color: t.color, text: t.text,
+      tColor: t.tColor, size, tX: t.x + boxWidth/2, tY: start + boxHeight/2 + 6, action: t.action,
     }
   })
 }